Rename macros.rs to utils.rs
